Return packed-query as bag of bytes (string)
This change looks pretty bonkers in isolation, so some context is
warranted.

Summary
-------
Preparation for packing multiple kinds of (sub)tasks into a single bag
of bytes.

Context
-------
The higher-level goal is to support attributes (and possibly other extra
data), which are per-trace or per-sample values such as coordinates,
which should be available in the data package retrievable at /result.

The problem is that the querying attributes might not be
straight-forward, as attributes for large cubes are many, and for
correctness it would be nice to re-use the geometry-aware query features
of oneseismic. When squinting, querying attributes is just another task,
which follows the same shape as the "main" query.

That means the definition of a task-set must be extended to support
arbitrary many "extra" queries other than the pure volume data. In a
way, this was well supported already in the data given by
mkschedule(), since it returns a vector, but some experiments made it
clear that it was pretty unergonomic to work with when there were
multiple "output tasks" that had to be packed.

To combat this, the scheduler outputs a bag-of-bytes, which is awkward
from inside C++, but (to my surprise) was slightly nicer when making the
bytes available to go. The new packing is a bit longer, but with fewer
failure points and more plainly work with bytes.

Interestingly, the approach could be replaced with

     result := bytes.Split(tasks, []byte("\0"))

but this will break down if the packing goes from being json/string
based into a binary format, and that complexity is better to keep in
C++. I anticipate that the go <-> C++ communication in
scheduler/MakeQuery will be revised soon enough.

All-in-all this commit remove a couple of failure points, and prepares
the schedule_maker for a beefier query definition. On the face of it
looks like a silly micro-optimisation, and while it allocates less, it
achieves something more.