Add 'attributes' concept to query and tasks
Introduce the 'attributes' field (in singular for individual tasks) that
are meant to control what extra, optional data oneseismic should query
and include in the response.

This commit only adds the fields to the messages. The goal is to support
queries in the spirit of:

    {
        cube(id: ...) {
            sliceByIndex(..., opts: {
                attributes: [cdp, utm]
            })
        }
    }

And the result should include cdp and utm coordinates for the slice.