Issue tasks for fetching slice attributes
This patch modifies the interface of schedule_maker slightly to support
multiple output tasks. Adding attributes to queries then boils down to
issuing tasks that are mostly indstinguishable to data queries to
downstream workers. This means most of oneseismics infrastructure (and
improvements!) can be leveraged when fulfilling queries for attributes.

This patch only adds support for attributes in slice queries, as it is
the simpler one. Since attribute support is a wide-spanning task, it
makes sense to get some good experience with the approach in the
simplest case (the slice) before extending support to all queries.

At a high level this doesn't change much, as downstream workers are
oblivious to what data is being queried. In the planner, some helpers
are pulled out and:

1. build() returns a vector-of-tasks, up from a single task
2. partition() traverses the vector-of-tasks and builds the packed task
   set