Add attributes option in slice query
Add the input type Opts, which is intended for setting optional
parameters such as the (now-implemented) attributes field. Attributes
is a generalisation of per-trace values such as coordinates.

The attributes is a narrow set of support values, which goes through a
normalize-and-unique filter that expands concepts into directional
attributes e.g. cdp -> cdpx, cdpy.

With the exception of enum-values (which is handled by graphql), all
model validation happens in C++, and the opts type is just forwarded
into the C++ query planner.