Add rough attribute (cdp only) upload
Add rough support for uploading cdpx and cdpy attributes. No command
line params or other knobs are added. The approach leverages the idea
that attributes are just the volume seen in 2D "from above", i.e. with a
very thin z axis. This enables re-use of virtually all infrastructure in
oneseismic.

The only attribute supported in the upload is cdp, which is arbitrary. I
wanted something that I could test and flow through the system, but did
not want the complexity of multiple kinds of attributes. It should be
easy enough to add support for more.