Remove broken python client tests
With the shutdown of the /query endpoint and the move to graphql, these
tests are completely broken. They were always quite high-maintenance
under churn, and have for a long time needed a better testing strategy.
There is some room for some re-inventing parts of the client too, which
would focus the tests more (e.g. msgpack -> xarray is a suite on its
own).

Rather than keeping broken tests around and have CI fail, just remove
the test and make space for a new system in its place.