Return sizes separately in planner
Drop the idea of using a single array (std::string) with individual
task-size in-band through structure, and introduce the taskset type.
In-band signalling separation of tasks with \0 worked well enough, but
restricts the task types to text protocols.

The new task-set is even slightly more ergonomic than using std::string,
and manages to bring some custom code with it, at the cost of an extra
vector.

Overall this is a significant clarity win, at the expense of some tiny
custom container class in the task-set.