Store process-header as msgpack, no round-trip
Rather than round-tripping the process header between C++ and Go,
produce it as msgpack directly from C++ and store the bytes.

When the ResultHeader was introduced it was assumed that go code wanted
to keep tabs on the response header, and that it would be sufficiently
different from the process header. As of now that has not really
manifested itself, and process management should probably take other
forms anyway.

This saves the transformation of essentially the same object:
process_header -> ProcessHeader -> ResultHeader -> client-unpack by
going directly process_header -> client-unpack. With parsing written in
C++, the same object can even be re-used for automated parsing.

ntasks is renamed to nbundles, as the message now is intended for
clients, not the server-internal scheduling.