Pack process-header in envelope in C++
Have go treat the process header as a bag of bytes (for the most part),
and wrap the process header in the envelope in C++. This strengthens
C++'s control of the model, and makes go cares less about the data
layout and more about being just an I/O layer.

This patch adds a msgpack dependency, but the msgpack C++ library is header
only, so there is no cmake entry.