Strip ProcessHeader struct in go
Remove unused members from the ProcessHeader struct. The process header
is sent to the client (to help assembly), and only inspected in go to
determine how many tasks to wait for.

When concepts like process table are introduced, the need might warrant
bringing some of these (or other) fields back.