Use a flat structure for index in process-header
Rather than nesting array-of-array (or vector-of-vector in C++), pass
the index as a flat array. This reduces message overhead slightly, and
makes parsing nicer in many cases.

The shape field became a casualty, as it was essentially unused. Its
intention was a hint to parsing routines to allocate as much space as in
the source, to support read-all extraction, but this turns out is pretty
achievable without the explicit field. The shape concept may be
reintroduced later, possibly in a different form.