Pack tiles as tuples, not maps
Pack the slice response, tiles, as a sequence of tuples rather than a
sequence of maps. By using tuples the message size is dramatically
reduced. I measured slices from a 500M cube in all dimensions [1]:

|             | dim 0 | dim 1 | dim 2 |
|-------------|-------|-------|-------|
| array size  | 2.45M | 0.68M | 0.58M |
| old payload | 6.21M | 2.07M | 1.78M |
| new payload | 3.45M | 1.15  | 0.99M |

This saves 40-50% on message size, which translates to less network
pressure and faster parsing. This comes with little code complexity, but
messages are less self-descriptive and rely on a shared understanding of
the schema (which was always the case anyway).

[1] New also includes the flattened index, but that should reduce by an
    additional 5-8 bytes or so