Pack curtain response as flat struct-of-arrays
Re-design the curtain response format as a struct-of-arrays, to make
parsing easier and faster.

When it comes to easier, the message itself guides extraction so much
that there is no need for the xyindex, and the header.index can be
plugged directly into xarray. Packing uses (array) slots rather than
maps which makes the messages leaner.

The new structure is slightly less self descriptive since there no
longer are explicit field names, but implicitly resolved through
position. This means a slightly looser coupling between fields in
different systems, but more intimate knowledge. Overall it is a win
since the fields are pretty static anyway.

Implementation wise it is pretty straight forward:

    1. packing uses msgpack and not nlohmann/json.
    2. _traces is renamed to _bundle
    3. the process header index no longer maps lineno -> cartesian,
       since the cartesian really only is useful when getting fragments
       for storage
    4. mapping parameters are laid out in arrays rather than structs,
       but from a high level follows the same structure.

This patch roughly halves the size of the messages I tried (comparing
with a regular slice request):

     OLD
     response size: 1.699053 mb
     array size: 0.6834 mb

     NEW
     response size: 0.8603 mb
     array size: 0.6834 mb