Include response message type tag
Including a message type tag in the header enables consumers to pick the
correct assembly function when the message is received, not a-priori.
This means decoders don't need to be tightly controlled by the code
paths that makes requests, but can instead function as a worker that
receives messages and parses them.

On top of that, a reusable assembler system means that buffers can be
reused when we get around to optimising for that.