Look up correct attribute-shape
When reading the attrshape for the in/crossline slices, the attrlabels
could get unpacked from a list-of-one to the actual string 'inline'.
Since strings are sequences, len('inline') works, which made the
attribute shape bonkers.

The python extraction is on borrowed time anyway, so just work around it
by explicitly getting index-zero (which is what it effectively did
anyway).