Add C++ decoder for query results
Introduce the decoder, a feature to parse and extract responses, the
binary-encoded messages for user consumption.

Why
---
Up until now there has really been only one client, the python one,
which leverages the msgpack library and extracts nicely into numpy and
xarray. While there can perfectly well be multiple decoders, a
centralised C++ one is a design choice:

1. There is no already separate message format specification language;
   it is inferred from implementation. Now this blessed implementation
   is in C++.
2. It can be re-used in virtually any language, so that oneseismic can
   be consumed in more contexts. While it is not so bad to write
   extraction in Python, Javascript is a different story.
3. Single source of bugs and fixes.
4. Single source of features, such as streaming support.

Design
------
Clients are reasonably aware that the body/bundles of data can arrive in
a arbitrary order. The decoder is designed so that it can buffer and
parse data as it arrives on the wire, making message processing
completed virtually the second the download (which is a significant)
bottleneck is done. Users are expected to "attach" an decoder onto a
download stream.

A fundamental concept of the message design is that it is valid
messagepack. The decoder does both messagepack parsing _and_ extraction
in one go. This is a design choice to improve performance by discarding
data once it has been processed, rather than keeping the full
messagepack structure around until it is done.

Final remarks
-------------
A C++ based decoder comes at some complexity, but with some significant
benefits. The big problems in this space really isn't the C++, but in
the message design (and its consequences).