Add github actions job for building wheels
One of the big drawbacks of depending on compiled code for python is
that wheels are *much* more work. When the library was python-ony, a
simple setup.py bdist_wheel on my dev machine was sufficient to build
a wheel that anyone could run, but no more.

Leverage github actions and the cibuildwheel module to build wheels as a
part of the CI pipeline. There is no automated wheel upload (for now;
wheels can be downloaded and re-uploaded with twine), but the wheel
building doubles as test case.

Wheels are built linux, mac, and windows, in both 32 and 64-bit
flavours for all supported python versions.

The build pipeline is in pyproject.toml rather than in the wheels.yml
file, since it enables running cibuildwheel locally. From the
cibuildwheel docs:

    Static configuration works across all CI systems, and can be used
    locally if you run cibuildwheel --plat linux. This is preferred, but
    environment variables are better if you need to change per-matrix
    element (CIBW_BUILD is often in this category, for example), or if
    you cannot or do not want to change a pyproject.toml file. You can
    specify a different file to use with --config-file on the command
    line, as well.