Inline definition of to_vec into append_vector_ids
This is just a helper function to convert from strongly-typed tuples
(FS, CS etc.) to vectors (for serialization), which at some point should
be removed in favour of serializing the tuples directly. MSVC seems to
get confused by the template type in the lambda, but it doesn't really
hurt readability to inline the function.