Make query return type Promise optional
Using optional return values makes for a less brittle API, since even
promises could be just a small part of a larger query. This has no
python client support as of now (hence the awkward for loop for
extraction), but the API might as well support it.

For a motivating example, consider this query:

    cube(ids: $ids) {
        year
        sliceByLineno(dim: $dim, lineno: $lineno)
    }

IDs would be a set of independent surveys in the same area (e.g. North
Sea), which would return a specific line across a large area from the
candidate surveys. If a line is not a part of the survey then it is
simply skipped. For good measure the year is included to weed out very
old surveys. This would not work if the query would error after the
first line without a survey.

Applications, such as the python sdk, might want to ensure that errors
are raised when the query could not be fully resolved, but the API would
be right in returning what it could resolve and null the remaining
fields.