Don't do blocking read on failure channel
This is a bug in the /result that's been there since it was written. It
accidentally worked when the failure channel was closed in collectResult
(which it shouldn't be).

err, ok := <-failure will block until there is a message on the channel
*or* it is closed, but its intention is to read a message if available,
otherwise just continue.

Replace it with a select { case ...; default } which actually has this
behaviour.