Make Promise a scalar type
Change the Promise from a struct { url; key; } to a scalar type. From
the GraphQL docs:

    A GraphQL object type has a name and fields, but at some point those
    fields have to resolve to some concrete data. That's where the
    scalar types come in: they represent the leaves of the query.

Scalars have application specific serialization, parsing, and
validation. The promise type is a pure output type in oneseismic, and
clients are expected to know that its content (which are unchanged, at
least for now).

This makes querying less annoying. The content of a Promise is useless
unless you have all the fields (url, key), so it makes no sense having
to specify them all. It transforms queries like this:

    # before
    cube(id: $id) {
        sliceByIndex(dim: $dim, index: $index) {
            key
            url
        }
    }

    # after
    cube(id: $id) {
        sliceByIndex(dim: $dim, index: $index)
    }