Remove the cubes/list-cubes query
Being able to list cubes has been on the chopping block for a while; it
is fundamentally broken permissions wise, quite slow, and not really all
that useful.

The list-cubes query was introduced with a developer need in mind; list
the cubes and grab a random ID to run tests on. For end-users, the guids
you care about are already stored somewhere else, and other ids are
completely meaningless; what end-users want is a proper catalogue.

Then there's the problem of permissions; the list-cubes call (in Azure)
requires access to List Containers, and there is no granularity; either
all containers can be listed or none can, so users will see a lot of IDs
they can't really access.

A proper catalogue backed by a real database (probably) should take the
cubes query's place at some point, but for now it is just removed.

So what's the end-user impact?

1. oneseismic ls no longer works
2. for guid in guids: ... no longer works
3. cube = cli.cubes[guid] no longer raises KeyError when the ID is bad,
   but fails when data is requested instead.

1 and 2 should not matter much, as these features were hardly used. As
for 3:

    >>> cube = cubes['bad-guild'] # no longer fails
    >>> cube.slice(dim = 0, lineno = lineno)
    Exception: {'message': 'Not found', 'path': ['cube']}

which is acceptable.