Invalidate on-behalf tokens close to expiration
Evict on-behalf-of (obo) tokens if close (<5min) to expiration. This
makes sure on-behalf-tokens are sufficiently new for all other code to
assume that if a blob fetch fails, it's not an expired token but either
a bad URL or insufficient permissions.

While still imperfect, this is pretty huge since it means that the
retry-once-if-failed approach can go away, and it is now ergonomic to
fetch tokens rather than bury the work in helpers. This is significant
because the goal is eventually to support SAS-based authorization and a
general authorize(url, ...) function.

Ideally there should be some solid tests for this functionality, but as
of now the entire storage abstraction layer is in solid flux, and
SAS+no-auth support should be in place as they will surely challenge any
assumptions made now.