Get obo-tokens without retry
With the cache being aware of the token expiration time, the need for
WithOnbehalfAndRetry goes away. This function served to handle all cases
of expired, bad resource and bad permissions, but is not precise enough
to do it well:

* if the permissions are bad, getting a new token will not help
* if the target is bad, getting a new token will not help
* expired-but-not-evicted and revoked tokens can be removed manually

This is also a change in preparation of a more general
"add-appropriate-authorization" layer, which means that the complexity
of the established solution must be reduced.