Use line numbers in curtain/head.index
In 61e43638f06319c4ddcb6f6cdd3cb15dfb5978ee, when curtainByIndex support
was added, the to_cartesian_inplace was moved to the from_json so that
the header building was consistent with the cartesian grid, but the
header() function was never updated to map the index back into line
numbers.

This patch fixes that bug, and brings the index back into in/crossline
numbers.