Group traces by fragment, fetch curtain correctly
It turns out that curtain has been broken pretty much since it was
introduced. It has coincidentally worked ok for the tests that I have
manually run, but the general feature has been broken. I suppose it is
an excellent case study in why you want good test support, but for
oneseismic to support thorough testing there needs to be some
refactoring, which is a patch for later.

The problem
-----------
So what's the problem? Querying curtains is conceptually simple - ask
for a set of (x, y) pairs, and get the the traces for those pairs.
Oneseismic does not restrict the x, y pairs, but usually they would
follow some interesting structure like a well path. It's is an
arbitrarily-shaped vertical intersection.

So far, so good. The problem is that oneseismic stores the data in
fragments, and *really* does not want to fetch fragments more than once
[1]. To attack this oneseismic maps the query to "buckets" (from hash
map nomenclature), i.e. traces that belong to the same fragment column
are grouped together, so that each fragment only has to be fetched once.
The *input query* is generally not aware of this (and it depends on how
the volume is fragment).

The grouping is not only important for the oneseismic internals, it is
also important for decoding. In order to not tag every trace segment
with (x,y,z) it uses its major/minor system (really just a product)
which relies on the locality of consecutive traces.

The bug
-------
In short, the planner and process/fetch assumed this grouping was done,
but no code ever did. My manual testing [2] happened to format the query
this way.

The fix
-------
The fix is really just to group, but just reading the code with no
context could be hard to follow. From a high-level it is just a sort
really - first, group in buckets based on the fragment-ID, then sort the
buckets internally. Both sorts are lexicographical.

Intended consequence
--------------------
The effect of this is that the output no longer is sorted
lexicographically by inline, crossline, but a (total) lexicographically
sorted output by trace ID was always broken. Clients are free to add
this behaviour, but from oneseismic's point of view clients must always
take care and check with the head.index when positioning traces.

This is from a small test program that gets a few traces in no
particular order with the curtainByLineno and this patch applied:

    Coordinates:
        inline     (x, y) int64 9973 9981 10009 9977 10057 10195 10295 10259
        crossline  (x, y) int64 2074 2064 2286 2600 2583 2534 2141 2157
      * time       (time) int64 0 4000 8000 12000 ... 3388000 3392000 3396000

[1] You could argue that some redundant fetches between workers could be
    good, in particular with caching in the mix, but the fetch-blob
    operation is the single most expensive operation in the system, and
    minimising the number of fetches goes a long way for performance and
    cost control.
[2] The test is the slice-curtain equivalence; request a set of traces
    that happen to be the same as what sliceBy* would resolve, and see
    if you get the same data.