Break up schedule_maker into free functions
Break up the scheduler_maker system into free overloaded functions. It
turns out that the generic algorithms can just as easily be written with
free templated functions, and the lookup rules through a class really
isn't used. Maybe shared-state-between-<type> will be useful at some
point in the future, but then that can rather be special-cased rather
than introducing heavy template machinery for the general case. This
benefit is a simplification.

All changes are really just in signatures; the schedule() and
partition() functions are free, and uses overloading to select the
correct build() and header() functions, rather than the class. Since the
selection was always derived only from the input- and output types then
there should be no real loss in flexibility.

The partition() and schedule() functions can still be considered
customisation points, but rather than having to go through the class it
is sufficient to provide a more-specific version.