Pop graphql params; store query params with task
Store and transmit the (user) URL query parameters with the task. This
is a step to support user-configurable blob queries, e.g. for shared
access signature authorization.

The ?query parsing in graphql/Get removes the parameters it reads and
cares about, as it does not really make sense to forward them to the
blob store.

This patch only puts the query params in the messages, it does not
actually inspect it or attempt to send to the blob store.