Refactor blob- and token-access to separate layers/modules.
Blob- and Token-access is now through interfaces with
POC-implementations for Azure and local files. The interfaces can
probably be refined and made more elegant - this is a first-shot
straightforward solution.

The purpose is to isolate implementation-details for storage from
service-code/business-logic, thus making it practical to replace the
storage-mechanism without re-writing other parts of the code. This is
particularly useful for testing, but also facilitates replacing Azure if
this becomes a goal.

Some integration-tests with sample-data are also provided to demonstrate
how testing can utilise this isolation. These tests focus on success or
failure for queries, as opposed to verifying data. Thus the tests define
behaviour of the GraphQL-endpoint and (to a large extent) ignores the
actual data returned by the endpoint.

System-tests using real Azure can be fused with (or based on) these
integration-tests to verify the Azure back-end. Obviously, test-data and
more tests can be added to further verify data.