Include curtain attr-height in message
Include the height/zlength in the curtain message, rather than assuming
it can be read from the index. This makes the decoding function more
obviousy and immediate (since it no longer needs to read the index), but
bloats the message with another byte or two per bundle (which I think
is an acceptable trade off).

On its own this change only makes the decoding slightly simpler, but
this opens up for decoding curtain attributes of non-trace-length in the
z direction.