Add attribute support in curtainBy queries
Introduce the opts argument to the curtainBy* family of queries, and
support getting attributes along with data for the curtain query.

The implementation is pretty straight forward for the go bits; add opts
to the schema, and include the opts parameter in the message.Query.

The C++/query planning follows the pattern laid out by the attribute
support in slice, and with the exception of some repetetive code it
should not be too surprising; it is like when getting describing traces,
except tasks are not issued for the z-dimension.

This implements server-side support for attributes, but the attributes
are not made available in client code, since there is no good interface
support for it yet. Still, from my test setup I get this for emulating
a slice with the curtain query:

    <xarray.DataArray 'curtain' (x, y: 201, time: 850)>
    array([[0., 0., 0., ..., 0., 0., 0.],
           [0., 0., 0., ..., 0., 0., 0.],
           [0., 0., 0., ..., 0., 0., 0.],
           ...,
           [0., 0., 0., ..., 0., 0., 0.],
           [0., 0., 0., ..., 0., 0., 0.],
           [0., 0., 0., ..., 0., 0., 0.]], dtype=float32)
    Coordinates:
        inline     (x, y) int64 9961 9963 9965 9967 9969 ... 10355 10357 10359 10361
        crossline  (x, y) int64 2321 2321 2321 2321 2321 ... 2321 2321 2321 2321
      * time       (time) int64 0 4000 8000 12000 ... 3388000 3392000 3396000
        cdpx       (x, y) float32 9.566e+06 9.566e+06 ... 9.686e+06 9.687e+06
        cdpy       (x, y) float32 1.011e+07 1.011e+07 ... 1.059e+07 1.059e+07