Revert "Find and link msgpack with cmake"
This reverts commit 8822e41ccdef5ea368e52f015e0eb06e561ba7f1.

This change is pretty broken, as it relies on newer versions of
libmsgpack [1], and was not thoroughly tested. It was doubly broken
since it did not actually link, but rather used the include-dir, and it
would not work for decoder-only builds.

Since we use the header-only feature relying on it being in an available
include path is good enough for now, although this bit of the makefiles
should probably be improved upon.

It was merged to make the macos CI build work again, but in this
particular broken instance the real fix was installing the msgpack-cxx
package; see commit 38487dc38f9229324eec741ab3a76b09ecaa3a60

[1] e.g. the target msgpackc-cxx is not found on my debian buster
    (which provides 3.0.1-3)