Use fixed-size ID in message/single
Fix the fragment ID tuple dimension to 3. While the geometry module
supports arbitrary dimensions, oneseismic as a system does not, and
this goes a long way to mark a subsystem for "not ready for ND".

Pinning to 3 dimension does not come without benefits; short-vector
optimisations are generally not available, which means we save a bunch
of heap allocs this way, and we get some obvious asserts.