Refactor curtain planner to single code path
Rather than effectively repeating the curtain planner code for
attributes, generalise it (with the lambda) so that the
"outer" structure is clearly derived from data + attributes. This does
introduce some extra syntax and an almost indirection for the proper
algorithm, but disconnects the transformation of traces into tasks from
the kind of task.

While the repetition itself was not so bad (only a handful of lines),
this approach clearly states that extracting attributes is not really
different from extracting data.