Merge column expansion and coordinates fill
Merge the two phases of the curtain/build algorithm into one. Rather
than first generating the "buckets" (fragments) in the first phase and
then looking up the same buckets to populate them with coordinates in
the second phase, do it in one go.

This means a full loop is gone, which in addition to making the
algorithm easier to understand should improve performance since the
input is not traversed twice, and there are fewer ID lookups.