Move lambda into tasks-loop body
Simplify the building function by changing the order of steps in the
algorithm to first create the task descriptions, then populating them.
This means the lambda stuff can go away in favour of a for-each-task
loop.

The overall complexity of the function does not change significantly,
but the syntactic elements are slightly less exotic.