Reimagine python support; separate user, dev APIs
Why
---
The established python end-user support (client.new(); cube =
client[guid] etc) suffers from being too biased by the old REST-y design
and trying to paper over some of some complexities. The end result of
that was a client that was difficult(ish) for users to use, and
difficult for developers to modify to take advantage of the new and
powerful query engine.

Rather than putting lipstick on a pig, this patch takes a step back and
re-evalulates what python support should look like. These are the
highlights:

* A new simple_client which unapologetically limits power and
  configurability to be easy to get started with.
* Free functions for typical developer tasks; process representation,
  URL generation, filtering. These never do I/O.
* The developer APIs generally only set up parameters, and are built
  on the idea that callers should pick their own graphql and http
  transport libs.
* Established boundaries between developer APIs for writing rich clients
  and libraries and direct-end-users for small programs and one-off
  solutions.

Developer APIs
--------------
The developer APIs (for now mostly decoding/ and internal/) never really
deal with building queries and doing transport. The idea is to offload
that work to 3rd party graphql clients, which can even bring their own
DSL. This means library devs can construct and perform their queries as
they see fit, without being restricted by whatever we choose to support
and (poorly) test. This comes from the notion that oneseismic never
really did anything before the query was constructed anyway, and
obfuscating a 3rd party got us nowhere.

Sophisticated programs that use oneseismic will usually have specific
needs for authentication, transport scheduling, etc., and this approach
gives them the most flexibility. Briefly, their usage would look
something like this:

    query = (cube(...)
        .select(sliceByIndex(...))
        .select(id)
        .select(linenumbers)
        .select(filename)
    )
    sas = from_cache(guid)
    with transport_with_sas(gqlclient, sas) as gc:
        r = gqlclient.execute(query)

    r = procs_from_promises(r.content)
    report_metadata(r)
    compute_on_slice(next(filter_procs(r)))

The simple module
-----------------
The simple module tries to be the stark contrast to the developer APIs.
With reasonable defaults, a large part of user programs should be fine
with a restricted set of queries that feel "python native". They don't
need the power of arbitrary queries, and certainly don't want to muck
about with transport objects and the likes.

This gives a *great* out-of-the-box user experience, and an easy way to
get started.

    import oneseismic.simple as simple
    sc = simple.simple_client(host)

    procs = []
    for dim in [0, 1, 2]:
        proc = sc.sliceByIndex(guid, dim, 150, attributes = ['cdp'])(sas = sas)
        procs.append(proc)

    for proc in procs:
        print(proc.xarray())