Ensure slice index for attrs is inside volume
Attributes are stored as a thin 3D volume with zdim = 1 (yields the top
plane) to fit nicely in the rest of the oneseismic infrastructure. When
querying attributes along with a z-slice (for any z), make sure the
index must be set to zero. idx % cube_shape[dim] is the identity for
unaltered volumes, since index < cube_shape[dim].