Kill on-behalf-of auth support
Remove support for on-behalf-of tokens as they don't really give us what
we need them to.

The OBO token support has accounted for many many lines of code in
oneseismic, and been a source of problems. Getting fresh tokens is very
slow (400ms+), which forces oneseismic to cache them. They require care
since we need to look them up from some pool rather than use what is
available directly in gin. They mandate custom oneseimic login and a
bunch of signing logic.

Then it turns out they don't work for service principals [1]. Since
oneseismic is designed to serve other systems as a database, and OBO
does not work for it, the better decision is to remove all OBO support
(at least for now) and focus on simpler auth mechanisms which work
smoothly for end users and services.

With the OBO gone, the need for oneseismic login is gone.

[1] https://docs.microsoft.com/en-us/azure/active-directory/develop/v2-oauth2-on-behalf-of-flow