Use dedicated type for 401, 403 manifest errors
Use a new, dedicated type for permission errors when getting the
manifest. This is the first auth check in oneseismic, and if this passes
then it is extremely unlikely for the request as a whole to fail (due to
permissions). The permission error type can be used to distinguish it
from other temporary or input issues, but as of now only serve to remove
the use of naked errors.New() calls.