Add queryContext, fix resolver parameters in type
Rather than using a map str -> str to carry parameters to the resolvers,
use a struct with named fields. This opens for heterogenous fields, an
explicit (and limited) key set, and eventually behaviour and automation.

On its own, this change is a no-op - it is only intended to make way for
more changes in the future, and provide a little bit of clarity.