Build url.URL object outside of getManifest
Make the getManifest helper take fully hydrated objects, rather than
having to parse URLs and guess what to do when the parsing fails.

This is a fairly low-value change in isolation as it only moves the
check to external call site, but it makes the getManifest less magic and
makes the relationship between the endpoint and args.Id stuff more
obvious where it matters (in the resolver).