Add session.QueryManifest()
Add a function for querying the session object using JSON pointer. The
goal is to centralise the manifest knowledge and error handling, and
transform go into a pure I/O layer.

A casualty of this work, with the introduction of query_result, is the
naked plan cleanup(). It has been renamed plan_delete to be consistent
with the query_result type and functions.