Add more elaborate log for failed session.Init()
A broken session.InitWithManifest() is really a non-recoverable
operation as it only happens when 1. a manifest exists and 2. it
is not parsable or convertable to a C++ manifest object. It can in
theory fail due to memory constraints (std::bad_alloc), but this is
probably even more rare, and must be handled the same way.