Add test for cube.Linenumbers
Add a test for the cube.Linenumbers resolver. The test is subtly
valuable, as the communication between go and C++ can be a bit wonky,
and is subject to change.

The resolver is constructed directly, in order to not depend on going
through some I/O step where the manifest is fetched from the blob store
or a disk or memory device. This makes the test fast and reliable, at
the cost of having to manually setup the cube+context internals.