Merge cube.basic* into basicQuery
The dispatching of a structure query (slice, curtain) only differ in the
content of the message - the function type and the args+opts, which
makes the functions essentially duplicated. Merge the two bodies into a
generic basicSlice.

This might be a coincidental duplication, and if the curtain and slice
should need different implementations then they should surely be split
up again.