Merge bodies of manifest queries
Using the session.ManifestQuery API and unpacking the results is pretty
much identical regardless of what exactly the query is, so the
implementations can be merged with some go shenanigans.

This lowers cost of adding new manifest-lookup fields tremendeously, and
boils down to the boilerplate and change the path argument and type.

Passing an interface{} in order to unmarshal into the right type must be
supported by proper unit tests, since there is no compiler guarantees
for getting it right.